switch from manually constructing searchResults to displaying via turboframes and stimulus
I had trouble getting the automatic search on input change working with stimulus controllers... at first.
So, I decided to move on and display my search results by receiving json payload from my api and constructing my searchResults with the document model (a lot of document.createElement and el.classList.add .... it got ugly)

This was a really simple component, yet it still got really ugly when doing it like this. I start to think "hmm, this much be why all the javascript libraries are so popular." So, i decided to comeback to figuring out how to use turbo and stimulus.

It wasn't so bad afterall. In fact, the answer was right here all along: https://discuss.hotwired.dev/t/triggering-turbo-frame-with-js/1622. I had actually found this discussion and I skipped right to the answers before I read the original post. The original post, in my opinion, is the easiest way to implement this functionality with stimulus. It really doesn't even feel that hacky to me.

All in all, this was a great exercise.. I got to see how nasty the document model can become when you are creating html client side, then it forced me to come back and understand turbo much better

One good learning: turbo allows you to target other frames from a frame. this is with the target attr. However, if the target you put doesn't match a target, you'll get a confusing error message. watch out for that